requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configobj==5.0.6
dbus-python==1.2.4
decorator==4.2.1
gpg==1.13.1
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
nftables==0.1
perf==0.1
pygobject==3.28.3
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-linux-procfs==0.6.2
pyudev==0.21.0
PyYAML==5.4.1
requests==2.25.1
rpm==4.14.3
schedutils==0.6
selinux==2.9
six==1.11.0
slip==0.6.4
slip.dbus==0.6.4
syspurpose==1.27.16
urllib3==1.26.3
